+++ lang/en/settings.php
@@ -128,6 +128,11 @@ return [
         'three_month' => '3 Months (90 Days)',
         'tile_description' => 'Controls system-wide settings such as logo, maintenance banner and page sizes.',
         'title' => 'Application',
+        'toast_lifetime' => [
+            'custom' => 'Custom',
+            'description' => 'Time in seconds for which a notification is displayed',
+            'title' => 'Display duration of pop-up messages',
+        ],
         'two_weeks' => '2 Weeks (14 Days)',
         'two_years' => '2 Year (730 Days)',
         'user_settings' => 'User settings',
@@ -171,6 +176,11 @@ return [
         'view' => 'Detailed information for the role :name',
     ],
     'room_types' => [
+        'bbb_api' => [
+            'create_parameters' => 'Additional Create API parameters',
+            'create_parameters_description' => 'Specify as attribute-value pair (one per line, without spaces), e.g. webcamsOnlyForModerator=true',
+            'title' => 'BigBlueButton API',
+        ],
         'color' => 'Color',
         'custom_color' => 'Custom color',
         'default_room_settings' => [