+++ tests/Feature/api/v1/Room/RoomTypeTest.php
@@ -78,10 +78,6 @@ class RoomTypeTest extends TestCase
                     'name',
                     'id',
                     'model_name',
-                    'restrict',
-                    'updated_at',
-                    'max_participants',
-                    'max_duration',
                 ],
             ]])
             ->assertJsonCount(4, 'data');
@@ -96,10 +92,6 @@ class RoomTypeTest extends TestCase
                     'name',
                     'id',
                     'model_name',
-                    'restrict',
-                    'updated_at',
-                    'max_participants',
-                    'max_duration',
                 ],
             ]])
             ->assertJsonCount(4, 'data');
@@ -114,10 +106,6 @@ class RoomTypeTest extends TestCase
                     'name',
                     'id',
                     'model_name',
-                    'restrict',
-                    'updated_at',
-                    'max_participants',
-                    'max_duration',
                     'everyone_can_start_default',
                     'everyone_can_start_enforced',
                     'mute_on_start_default',
@@ -273,6 +261,7 @@ class RoomTypeTest extends TestCase
                 'record_attendance_enforced',
                 'has_access_code_default',
                 'has_access_code_enforced',
+                'create_parameters',
             ]]);
 
         // Test deleted
@@ -334,6 +323,7 @@ class RoomTypeTest extends TestCase
             'auto_start_recording_enforced' => false,
             'has_access_code_default' => true,
             'has_access_code_enforced' => false,
+            'create_parameters' => 'meetingLayout=PRESENTATION_FOCUS',
         ];
 
         // Test guests
@@ -397,6 +387,7 @@ class RoomTypeTest extends TestCase
                 'auto_start_recording_enforced' => false,
                 'has_access_code_default' => true,
                 'has_access_code_enforced' => false,
+                'create_parameters' => 'meetingLayout=PRESENTATION_FOCUS',
             ]);
 
         // Test with invalid data
@@ -444,6 +435,7 @@ class RoomTypeTest extends TestCase
             'auto_start_recording_enforced' => 'no',
             'has_access_code_default' => 'yes',
             'has_access_code_enforced' => 'no',
+            'create_parameters' => $this->faker->textWithLength(65001),
         ];
         $this->actingAs($this->user)->postJson(route('api.v1.roomTypes.store'), $data)
             ->assertJsonValidationErrors([
@@ -490,6 +482,7 @@ class RoomTypeTest extends TestCase
                 'auto_start_recording_enforced',
                 'has_access_code_default',
                 'has_access_code_enforced',
+                'create_parameters',
             ]);
 
         $data['roles'] = [1337];
@@ -601,6 +594,7 @@ class RoomTypeTest extends TestCase
             'auto_start_recording_enforced' => false,
             'has_access_code_default' => true,
             'has_access_code_enforced' => false,
+            'create_parameters' => 'meetingLayout=PRESENTATION_FOCUS',
         ];
 
         // Test guests
@@ -670,6 +664,7 @@ class RoomTypeTest extends TestCase
                 'auto_start_recording_enforced' => false,
                 'has_access_code_default' => true,
                 'has_access_code_enforced' => false,
+                'create_parameters' => 'meetingLayout=PRESENTATION_FOCUS',
             ]);
 
         $roomType->refresh();
@@ -730,6 +725,7 @@ class RoomTypeTest extends TestCase
             'auto_start_recording_enforced' => 'no',
             'has_access_code_default' => 'yes',
             'has_access_code_enforced' => 'no',
+            'create_parameters' => $this->faker->textWithLength(65001),
             'updated_at' => $roomType->updated_at,
         ];
         $this->actingAs($this->user)->putJson(route('api.v1.roomTypes.update', ['roomType' => $roomType->id]), $data)